Make it work on Java 17
With these changes, it works on Java 17 as long as it has the following
JVM arguments:

  --add-opens=java.base/java.lang=ALL-UNNAMED --add-opens=java.base/jdk.internal.misc=ALL-UNNAMED

+++ src/main/java/net/intelie/introspective/reflect/StringFastPath.java
@@ -50,9 +50,9 @@ public class StringFastPath {
     public long size(String s) {
         if (isCompactStringsEnabled()) {
             byte coder = U.getByte(s, coderOffset);
-            return JVMPrimitives.align(stringSize) + JVMPrimitives.align((s.length() << coder) + arrayOffset);
+            return JVMPrimitives.align(stringSize) + JVMPrimitives.align(((long) s.length() << coder) + arrayOffset);
         } else {
-            return JVMPrimitives.align(stringSize) + JVMPrimitives.align(arrayOffset + s.length() * 2);
+            return JVMPrimitives.align(stringSize) + JVMPrimitives.align(arrayOffset + s.length() * 2L);
         }
     }
